Move {push,pop}_{macros,warnings}.h to fast_io_dsal/impl
