+++ include/fast_io_dsal/vector.h
@@ -4,8 +4,8 @@
 #error "You must be using a C++ compiler"
 #endif
 
-#include "../fast_io_core_impl/misc/push_macros.h"
-#include "../fast_io_core_impl/misc/push_warnings.h"
+#include "impl/push_macros.h"
+#include "impl/push_warnings.h"
 
 #include <version>
 #include <type_traits>
@@ -63,5 +63,5 @@ using vector = ::fast_io::containers::vector<T, Alloc>;
 
 #endif
 
-#include "../fast_io_core_impl/misc/pop_macros.h"
-#include "../fast_io_core_impl/misc/pop_warnings.h"
+#include "impl/pop_macros.h"
+#include "impl/pop_warnings.h"
