+++ include/fast_io_freestanding.h
@@ -11,7 +11,7 @@
 
 #include "fast_io_core.h"
 
-#include "fast_io_core_impl/misc/push_macros.h"
+#include "fast_io_dsal/impl/push_macros.h"
 
 #include "fast_io_freestanding_impl/exception.h"
 // #include"fast_io_freestanding_impl/posix_error.h"
@@ -35,6 +35,6 @@
 #include "fast_io_unit/bitset.h"
 #endif
 
-#include "fast_io_core_impl/misc/pop_warnings.h"
+#include "fast_io_dsal/impl/pop_warnings.h"
 
 #endif
